Refactor: put all widgets in separate files Fixes #225
