+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:crowdleague/widgets/crowd_league_app.dart';
+import 'package:crowdleague/widgets/app/crowd_league_app.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(CrowdLeagueApp());
