+++ lib/widgets/auth/auth_page/buttons/twitter_sign_in_fab.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class TwitterSignInFAB extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Opacity(
+      opacity: 0.5,
+      child: FloatingActionButton(
+          child: ImageIcon(AssetImage('assets/images/twitter_logo.png')),
+          elevation: 1,
+          mini: true,
+          onPressed: null),
+    );
+  }
+}
