+++ lib/widgets/auth/auth_page/static_elements/crowd_league_logo.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class CrowdLeagueLogo extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Image(
+      image: AssetImage('assets/images/logo-300-greyscale.png'),
+      colorBlendMode: BlendMode.darken,
+      width: 150,
+      height: 150,
+    );
+  }
+}
