+++ lib/widgets/auth/auth_page/static_elements/tag_line_text.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class TaglineText extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Text(
+          'BE IN A LEAGUE',
+          style: TextStyle(fontSize: 20),
+        ),
+        Text(
+          'OF YOUR OWN',
+          style: TextStyle(fontSize: 20),
+        ),
+      ],
+    );
+  }
+}
