+++ lib/widgets/auth/email_auth_options_page/buttons/password_suffix_icon_button.dart
@@ -1,0 +1,22 @@
+import 'package:crowdleague/actions/auth/update_email_auth_options_page.dart';
+import 'package:crowdleague/extensions/extensions.dart';
+import 'package:flutter/material.dart';
+
+class PasswordSuffixIconButton extends StatelessWidget {
+  final bool visible;
+
+  const PasswordSuffixIconButton({Key key, this.visible}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsetsDirectional.only(end: 12.0),
+      child: IconButton(
+        icon: (visible) ? Icon(Icons.close) : Icon(Icons.remove_red_eye),
+        onPressed: () {
+          context.dispatch(UpdateEmailAuthOptionsPage(showPassword: !visible));
+        },
+      ),
+    );
+  }
+}
