+++ lib/widgets/auth/email_auth_options_page/buttons/sign_in_button.dart
@@ -1,0 +1,48 @@
+import 'package:crowdleague/actions/auth/sign_in_with_email.dart';
+import 'package:crowdleague/actions/auth/update_email_auth_options_page.dart';
+import 'package:crowdleague/enums/auto_validate.dart';
+import 'package:crowdleague/extensions/extensions.dart';
+import 'package:flutter/material.dart';
+
+class SignInButton extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return ButtonTheme(
+      height: 40.0,
+      padding: EdgeInsets.only(left: 30.0, right: 30.0),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(3.0),
+        side: BorderSide(
+          color: Colors.black,
+        ),
+      ),
+      child: RaisedButton(
+        onPressed: () {
+          if (Form.of(context).validate()) {
+            context.dispatch(SignInWithEmail());
+          } else if (!Form.of(context).validate()) {
+            // set form to autovalidate on user input
+            // validation would else occur only on form submit
+            context.dispatch(UpdateEmailAuthOptionsPage(
+              autovalidate: AutoValidate.onUserInteraction,
+            ));
+          }
+        },
+        color: Colors.white,
+        child: Row(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text(
+              'SIGN IN',
+              style: TextStyle(
+                  fontSize: 16.0,
+                  fontFamily: 'SF Pro',
+                  fontWeight: FontWeight.w500,
+                  color: Colors.black),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
