+++ lib/widgets/conversations/conversation_page/conversation_page.dart
@@ -1,8 +1,8 @@
 import 'package:crowdleague/actions/conversations/disregard_messages.dart';
 import 'package:crowdleague/actions/conversations/observe_messages.dart';
 import 'package:crowdleague/models/app/app_state.dart';
-import 'package:crowdleague/widgets/chats/messages/message_entry.dart';
-import 'package:crowdleague/widgets/chats/messages/messages_list.dart';
+import 'package:crowdleague/widgets/conversations/conversation_page/message_entry/message_entry.dart';
+import 'package:crowdleague/widgets/conversations/conversation_page/messages_list.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
@@ -26,12 +26,7 @@ class MessagesPage extends StatelessWidget {
               ? Center(child: CircularProgressIndicator())
               : Center(
                   child: Column(
-                  children: [
-                    Expanded(
-                      child: MessagesList(),
-                    ),
-                    MessageEntry()
-                  ],
+                  children: [Expanded(child: MessagesList()), MessageEntry()],
                 ));
         },
       ),
