+++ lib/widgets/conversations/conversation_page/message_entry/message_entry.dart
@@ -1,0 +1,15 @@
+import 'package:crowdleague/widgets/conversations/conversation_page/message_entry/message_text_field.dart';
+import 'package:crowdleague/widgets/conversations/conversation_page/message_entry/save_message_button.dart';
+import 'package:flutter/material.dart';
+
+class MessageEntry extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        MessageTextField(),
+        SaveMessageButton(),
+      ],
+    );
+  }
+}
