+++ lib/widgets/conversations/conversation_page/message_entry/message_text_field.dart
@@ -1,35 +1,6 @@
-import 'package:crowdleague/actions/conversations/save_message.dart';
 import 'package:crowdleague/actions/conversations/update_conversation_page.dart';
 import 'package:crowdleague/extensions/extensions.dart';
 import 'package:flutter/material.dart';
-
-class MessageEntry extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        MessageTextField(),
-        SaveMessageButton(),
-      ],
-    );
-  }
-}
-
-class SaveMessageButton extends StatelessWidget {
-  const SaveMessageButton({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 25),
-      child: IconButton(
-          icon: Icon(Icons.send),
-          onPressed: () => context.dispatch(SaveMessage())),
-    );
-  }
-}
 
 class MessageTextField extends StatelessWidget {
   @override
