+++ lib/widgets/conversations/conversation_summaries_page/conversation_summaries_page.dart
@@ -3,7 +3,7 @@ import 'package:crowdleague/actions/conversations/disregard_conversations.dart';
 import 'package:crowdleague/actions/conversations/observe_conversations.dart';
 import 'package:crowdleague/models/app/app_state.dart';
 import 'package:crowdleague/models/conversations/conversation_summary.dart';
-import 'package:crowdleague/widgets/chats/conversations/conversations_list.dart';
+import 'package:crowdleague/widgets/conversations/conversation_summaries_page/conversation_summaries_list.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
@@ -16,7 +16,7 @@ class ConversationSummariesPage extends StatelessWidget {
       distinct: true,
       converter: (store) => store.state.conversationSummariesPage.summaries,
       builder: (context, summaries) {
-        return ConversationsList(
+        return ConversationSummariesList(
           summaries: summaries,
         );
       },
