+++ lib/widgets/main/main_page_body.dart
@@ -1,0 +1,28 @@
+import 'package:crowdleague/enums/nav_bar_selection.dart';
+import 'package:crowdleague/widgets/conversations/conversation_summaries_page/conversation_summaries_page.dart';
+import 'package:crowdleague/widgets/more_options/more_options_page.dart';
+import 'package:flutter/material.dart';
+
+class MainPageBody extends StatelessWidget {
+  final NavBarSelection selection;
+  const MainPageBody({
+    @required this.selection,
+    Key key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    switch (selection) {
+      case NavBarSelection.home:
+        return Center(child: Text('Home Page'));
+      case NavBarSelection.business:
+        return Center(child: Text('Business Page'));
+      case NavBarSelection.conversations:
+        return ConversationSummariesPage();
+      case NavBarSelection.more:
+        return MoreOptionsPage();
+      default:
+        return Center(child: Text('Main Page'));
+    }
+  }
+}
