+++ lib/widgets/main/nav_bar.dart
@@ -1,0 +1,50 @@
+import 'package:crowdleague/actions/navigation/store_nav_bar_selection.dart';
+import 'package:crowdleague/enums/nav_bar_selection.dart';
+import 'package:crowdleague/extensions/extensions.dart';
+import 'package:flutter/material.dart';
+
+class NavBar extends StatelessWidget {
+  final int selectedIndex;
+
+  const NavBar({
+    @required this.selectedIndex,
+    Key key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BottomNavigationBar(
+      items: const <BottomNavigationBarItem>[
+        BottomNavigationBarItem(
+          icon: Text(
+            'sports_basketball',
+            style: TextStyle(
+              fontFamily: 'maticons',
+            ),
+          ),
+          label: 'Home',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.business),
+          label: 'Business',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.message),
+          label: 'Conversations',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.more_vert),
+          label: 'More',
+        ),
+      ],
+      type: BottomNavigationBarType.fixed,
+      currentIndex: selectedIndex,
+      onTap: (index) => _onItemTapped(context, index),
+    );
+  }
+
+  void _onItemTapped(BuildContext context, int index) {
+    context.dispatch(
+        StoreNavBarSelection(selection: NavBarSelection.valueOfIndex(index)));
+  }
+}
