+++ test/widget/auth/email_auth_options_page_widget_test.dart
@@ -6,7 +6,12 @@ import 'package:crowdleague/enums/auto_validate.dart';
 import 'package:crowdleague/enums/email_auth_mode.dart';
 import 'package:crowdleague/models/app/app_state.dart';
 import 'package:crowdleague/reducers/app_reducer.dart';
-import 'package:crowdleague/widgets/auth/email_auth_options_page.dart';
+import 'package:crowdleague/widgets/auth/email_auth_options_page/buttons/create_account_button.dart';
+import 'package:crowdleague/widgets/auth/email_auth_options_page/buttons/sign_in_button.dart';
+import 'package:crowdleague/widgets/auth/email_auth_options_page/email_auth_options_page.dart';
+import 'package:crowdleague/widgets/auth/email_auth_options_page/text_fields/email_text_field.dart';
+import 'package:crowdleague/widgets/auth/email_auth_options_page/text_fields/password_text_field.dart';
+import 'package:crowdleague/widgets/auth/email_auth_options_page/text_fields/repeat_password_text_field.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:flutter_test/flutter_test.dart';
