+++ test/widget/crowd_league_app/crowd_league_app_test.dart
@@ -7,7 +7,8 @@ import 'package:crowdleague/actions/notifications/print_fcm_token.dart';
 import 'package:crowdleague/actions/notifications/request_fcm_permissions.dart';
 import 'package:crowdleague/models/app/app_state.dart';
 import 'package:crowdleague/reducers/app_reducer.dart';
-import 'package:crowdleague/widgets/crowd_league_app.dart';
+import 'package:crowdleague/widgets/app/crowd_league_app.dart';
+import 'package:crowdleague/widgets/app/initializing_indicator.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:redux/redux.dart';
